% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_closest_distance.R
\name{average_closest_distance}
\alias{average_closest_distance}
\title{Calculate average closest network distance (d(c)) between drug targets and the nearest disease protein
Proven useful for drug efficacy scressning in Guney et al,  NatCom, 2016 (PMID: 26831545)}
\usage{
average_closest_distance(ppin_dist, from, to)
}
\arguments{
\item{ppin_dist}{Distances between all proteins included in the LCC of the PPIN.}

\item{from}{Vector specifying genes between which distances should be calculated. E.g.: from = disease_genes}

\item{to}{Vector specifying genes between which distances should be calculated. E.g.: to = drug_targets}
}
\value{
Double specifying the average closest network distance between elements in 'from' and elements in 'to'
}
\description{
Calculate average closest network distance (d(c)) between drug targets and the nearest disease protein
Proven useful for drug efficacy scressning in Guney et al,  NatCom, 2016 (PMID: 26831545)
}
