% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/average_closest_distance_network_drug_screening.R
\name{average_closest_distance_network_drug_screening}
\alias{average_closest_distance_network_drug_screening}
\title{Closest Distance Network Drug Screening}
\usage{
average_closest_distance_network_drug_screening(
  ppin,
  drug_target_matrix,
  disease_genes,
  file_name,
  disease_genes_lcc = F,
  min_bin_size = 100,
  n_random_iterations = 1000,
  cores = 1,
  prepare_recycle = F,
  recycle = F,
  out_dir = getwd(),
  seed = 35
)
}
\arguments{
\item{ppin}{Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{drug_target_matrix}{Matrix listing each drug's targets. One column per drug.}

\item{disease_genes}{Vector specifying disease associated genes (DEGs, GWAS genes, OMIM genes, etc.)}

\item{file_name}{Character string specifying filename for the output.}

\item{disease_genes_lcc}{Boolean, if TRUE the LCC of disease genes is used for the calculations instead of disease genes.}

\item{min_bin_size}{Integer. The bin-size used for node degree adjustment of network calculations.}

\item{n_random_iterations}{number of permutations used to derive reference distribution of network distances.}

\item{cores}{number of cores used for computing.}

\item{prepare_recycle}{Boolean, if TRUE ppin_distances, bins and a cleaned version of drug_target_matrix are saved in  "/temp" folder for shortening computational time in coming iterations.}

\item{recycle}{Boolean, if TRUE ppin_distances and bins are loaded from "/temp". Decreases computational time. Can not be used if parallel sessions are accessing the same folder.}

\item{out_dir}{Character string specifying path to out directory where data will be saved.}
}
\value{
Returns output matrix specifying network distances of the inputted drugs.
}
\description{
Used for calculations of closest network proximity. Memory heavy.
}
\examples{
load_example_data()
ppin <- ppin_formatting(lit_ppi)
drug_target_matrix <- prepare_drug_target_matrix_for_network_distance_calculation(ppin, drug_target_matrix, file_name = "in_lit_ppin", out_dir = "data/")
closest_distance_network_drug_screening(ppin = ppin, drug_target_matrix = drug_target_matrix, file_name = "test")

}
