% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_genes_NicheNet.R
\name{background_genes_NicheNet}
\alias{background_genes_NicheNet}
\title{Calculation of background genes for NicheNet ligand activity analysis}
\usage{
background_genes_NicheNet(data, cell_IDs, exp_cutoff = 0.2)
}
\arguments{
\item{data}{Matrix. DCA denoised, log10 transformed expression matrix. Gene names as row names. Cell IDs as column names.}

\item{cell_IDs}{Vector or list. Cell IDs for the cell population (e.g. the cluster / cell type) of interest. If list, every list object contains all cell labels for a given cell type.}

\item{exp_cutoff}{Integer or float. Expression cut-off to be considered a background gene. The standard setting corresponds to a mean log2 transformed gene expression of >= 0.1 among the cell population of interest. This cut-off was chosen to generate ca 10 000 background genes per cell population as recommended by the authors of NicheNet. The value might need to be adapted to your individual data set. The same cut-off should be applied to all cell populations within a data set.}
}
\value{
A vector of the length nrow(data) that specifies all genes passing the expression cut-off. If cell_IDs is a matrix, specifying all
}
\description{
Calculation of background genes for NicheNet ligand activity analysis
}
