% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_creation_by_min_bin_size.R
\name{bin_creation_by_min_bin_size}
\alias{bin_creation_by_min_bin_size}
\title{Bin creation for network degree adjustments}
\usage{
bin_creation_by_min_bin_size(ppin, min_bin_size)
}
\arguments{
\item{ppin}{Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{min_bin_size}{Integer. The bin-size used for node degree adjustment of network calculations.}
}
\value{
Returns matrix specifying the bins for further analysis.
}
\description{
Bin creation for network degree adjustments
}
