% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_evaluation_files.R
\name{combine_evaluation_files}
\alias{combine_evaluation_files}
\title{Combination of evaluation files for checking manualy if drugs counteract targeted DEGs fold change}
\usage{
combine_evaluation_files(files, label_for_files, output_file_name, cores = 1)
}
\arguments{
\item{files}{character vector; specifying full file names including paths for all evaluation files that should be combined into one output file.}

\item{label_for_files}{character vector; same length as 'files'. Specifies the label for each file that should be used in output file.}

\item{output_file_name}{character string; specifies name of output file (including output directory path and ".txt" ending)}

\item{cores}{number of cores used for computation.}
}
\value{
Saves a combined file in the chosen directory.
}
\description{
Combination of evaluation files for checking manualy if drugs counteract targeted DEGs fold change
}
