% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_drug_target_matrix.R
\name{create_drug_target_matrix}
\alias{create_drug_target_matrix}
\title{Closest Distance Network Drug Screening}
\usage{
create_drug_target_matrix(drugID, target)
}
\arguments{
\item{drugID}{Vector with IDs Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{target}{Vector with the same length as drugID that shall be sorted into different rows based on drugID.}
}
\value{
Matrix with length(unique(drugID)) columns. Each column specifies all targets for a certain drug. Column names specify which drugID the targets correspond to.
}
\description{
Used for calculations of closest network proximity. Memory heavy.
}
