% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_and_formatting.R
\name{data_analysis_and_formatting}
\alias{data_analysis_and_formatting}
\title{Data formatting of basic files used for network calculations}
\usage{
data_analysis_and_formatting()
}
\value{
Creates a drug target matrix based on DrugBank input
}
\description{
Data formatting of basic files used for network calculations
}
\examples{
data_analysis_and_formatting()

}
