% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_candidate_selection.R
\name{drug_candidate_selection}
\alias{drug_candidate_selection}
\title{Selection of drug candidates that fulfill network distance criteria dc < 1 and zc < -1.64 (that is P < 0.05 for network distance compared to random)}
\usage{
drug_candidate_selection(input)
}
\arguments{
\item{input}{Matrix. Column 4 = dc and column 8 = P value.}
}
\value{
Returns filtered matrix (excluding candidates that do not fulfill criteria). If all candidates are excluded this function returns NULL.
}
\description{
Selection of drug candidates that fulfill network distance criteria dc < 1 and zc < -1.64 (that is P < 0.05 for network distance compared to random)
}
\keyword{internal}
