% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_effect_on_targeted_degs.R
\name{drug_effect_on_targeted_degs}
\alias{drug_effect_on_targeted_degs}
\title{Title}
\usage{
drug_effect_on_targeted_degs(input, degs, drug_effect)
}
\arguments{
\item{input}{matrix with 8 columns in ```mode(degs) <- "character"```, listing individual drug's network distances. Column 1 = DrugBank ID (character string), column 2 = known drug for this disease (boolean), column 3 = n of drug targets (integer), column 4 = dc (double), column 5 = zc (double), column 6 = mean random dc (double), column 7 = SD of random dc (double), column 8 = P value for significance of dc (double). Only column 1 is used for merging information in this function, though the matrix is reshuffled before return and therefore requires all 8 columns.}

\item{degs}{matrix with 6 columns in ```mode(degs) <- "character"```, listing DEGs associated with the input file. Prserves output format of Seurat based DEG calculation. Column 1 = gene name in annotation that matches 'drug_effect', column 3 = average log fold change (avg_logFC), column 6 = adjusted P value.}

\item{drug_effect}{matrix with 3 columns in ```mode(degs) <- "character"```, listing drug target information derived from DrugBank. Column 1 = DrugBankID used for matching with 'input', column 2 = gene name in annotation that matches with 'degs', column 3 = pharmacological action on the drug target.}
}
\value{
Returns matrix for manual evaluation of drug effect on targeted DEGs. Matches all matrices and determins fold change of targeted DEGs.
}
\description{
Title
}
