% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LCC.R
\name{extract_LCC}
\alias{extract_LCC}
\title{Largest connected component}
\usage{
extract_LCC(network)
}
\arguments{
\item{network}{igraph object.}
}
\value{
Node labels of nodes included in LCC.
}
\description{
Extraction of the largest connected component (LCC) within an igraph object
}
\examples{
ppin <- as.data.frame(read.table(file = "data-raw/lit_ppi.txt", sep="\t", header = T, stringsAsFactors = F))
ppin_graph <- igraph::graph_from_data_frame(ppin, directed = F)
lcc_genes <- extract_LCC(ppin_graph) # only ppin proteins that are connected to other ppin proteins

}
