% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LCC_by_gene_set_of_interest.R
\name{extract_LCC_by_gene_set_of_interest}
\alias{extract_LCC_by_gene_set_of_interest}
\title{Extraction of the LCC formed by a gene set of interest in a PPIN}
\usage{
extract_LCC_by_gene_set_of_interest(network, gene_set_of_interest, plot = F)
}
\arguments{
\item{network}{igraph object. Largest connceted component of all proteins in a protein-protein interaction network (PPIN).}

\item{gene_set_of_interest}{Vector specifying nodes in 'network'}

\item{plot}{Boolean, if TRUE creates plot of depicturing isolated components formed by 'gene_set_of_interest' in 'network'}
}
\value{
Vector specifying the node names of the LCC formed by the 'gene_set_of_interest'.
}
\description{
Extraction of the LCC formed by a gene set of interest in a PPIN
}
\examples{
Checks for the LCC formed by 200 unique nodes in the network
ppin <- as.data.frame(read.table(file = "data-raw/lit_ppi.txt", sep="\t", header = T, stringsAsFactors = F))
goi <- unique(as.vector(ppin))[1:200]
ppin_graph <- igraph::graph_from_data_frame(ppin, directed = F)
lcc_genes <- extract_LCC_by_gene_set_of_interest(network = ppin_graph, gene_set_of_interest = goi) # only ppin proteins that are connected to other ppin proteins

}
