% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_evaluation_done.R
\docType{data}
\name{fc_evaluation_done}
\alias{fc_evaluation_done}
\title{Final file for the evaluation of the fold change criterion.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 59 rows and 20 columns.
}
\usage{
fc_evaluation_done
}
\description{
Gene annotation is Entrez gene symbols.
}
\keyword{datasets}
