% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_drug_prioritization.R
\name{final_drug_prioritization}
\alias{final_drug_prioritization}
\title{Final drug prioritization}
\usage{
final_drug_prioritization(
  fc_evaluation,
  pos_DrugID = 1,
  pos_clusterID = ncol(fc_evaluation),
  keep = c(pos_DrugID, pos_clusterID),
  inter_cent,
  intra_cent,
  file_name,
  out_dir = ""
)
}
\arguments{
\item{fc_evaluation}{Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{pos_DrugID}{Integer specifying the column in which the cluster IDs / cell type labels can be found that are to be matched with names in inter_cent.}

\item{pos_clusterID}{Integer specifying the column in which the cluster IDs / cell type labels can be found that are to be matched with names in inter_cent.}

\item{keep}{Vector specifying which columns (as indicated by column number) to keep in the final drug ranking file.}

\item{inter_cent}{Vector specifying the centrality score of each cluster. The name of each element indicates which cluster/cell type the centrality score belongs to. Names must be matchable with cluster identifiers in 'fc_evaluation'.}

\item{intra_cent}{Vector specifying the cluster-wide average of the intracellular centralities. Names indicate which drug the centrality score belongs to. Names must be matchable with column 1 in 'fc_evaluation'.}

\item{file_name}{Character string specyfing the file name by which the final ranking will be saved.}

\item{out_dir}{Character string specifying path to out directory where the final ranking will be saved.}
}
\value{
a matrix
}
\description{
Ranking of candidates passing selection criteria. Candidates are primarily ranked by the sum of the intercellular centrality scores of clusters in which they are considered candidates. Secondarily, candidates are ranked by their cluster-wide average intracellular centrality.
}
