% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_drug_centrality.R
\name{intracellular_drug_target_centrality}
\alias{intracellular_drug_target_centrality}
\title{Intracellular drug traget celtrality calculation}
\usage{
intracellular_drug_target_centrality(
  ppin,
  drug_target_matrix,
  degs,
  file_name = "",
  cores = 1,
  centrality_alg = "eigenvector centralities",
  out_dir = "",
  seed = 35
)
}
\arguments{
\item{ppin}{Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{drug_target_matrix}{Matrix listing each drug's targets. One column per drug. Only drug targets that are included in 'ppin' are entered used. If drug targets that are not included in 'ppin' are included they will be removed prior to centrality calculations.}

\item{degs}{Matrix or vector. If matrix, DEGs for each cell type are stored in a column.}

\item{file_name}{Character string specifying the first part of the filename for the output.}

\item{cores}{number of cores used for computing.}

\item{centrality_alg}{string. Specifying the centrality algorithm that shall be used for calculation of intracellular centrality. Standard is 'eigenvector centralities'. The function 'calculate_centralities()' from the CINNA package is used for calculation of centralities and hence this parametar should be set according to what is available for this function. Other valid options include (but are not limited to): "K-core Decomposition", "Kleinberg's hub centrality scores", "Laplacian Centrality", "Leverage Centrality", "Group Centrality", "Local Bridging Centrality".}

\item{out_dir}{Character string specifying path to out directory where data will be saved.}
}
\value{
a matrix including the geometric mean drug target centrality in every cell types LCC (if existent) as well as the arithmetric mean of all geometric mean centrality scores. The later of which was by us used for prioritization among drugs.
}
\description{
If no Largest Connected Component (LCC) can be found for a certain cell type, this cell type will be excluded completly from the calculations. As LCC we here defined a component of at least three proteins interacting with each other.
}
