% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_DCA.R
\docType{data}
\name{latent_DCA}
\alias{latent_DCA}
\title{Deep count autoencoder latent features}
\format{
A matrix with 32 latent features from DCA's latent space (rows) for 1500 cells (columns).
}
\usage{
latent_DCA
}
\description{
Deep count autoencoder latent features
}
\keyword{datasets}
