% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load scPred example data used for demonstrations of functions}
\usage{
load_example_data()
}
\value{
Several variables to the global environment
}
\description{
Load scPred example data used for demonstrations of functions
}
\examples{
load_example_data()

}
