% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load sample data
Downsized data sets for learning the structure and uses of the R package}
\usage{
load_sample_data()
}
\description{
Load sample data
Downsized data sets for learning the structure and uses of the R package
}
