% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genes.R
\docType{data}
\name{marker_genes}
\alias{marker_genes}
\title{Deep count autoencoder latent features}
\format{
A vector with 11 elements representing marker genes for the cell typing of the example data.
}
\usage{
marker_genes
}
\description{
Deep count autoencoder latent features
}
\keyword{datasets}
