% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheNet_cell_type_centrality.R
\name{NicheNet_cell_type_centrality}
\alias{NicheNet_cell_type_centrality}
\title{NicheNet ligand activity analysis for all cell types
Will test cell type centrality based on NicheNet ligand activities. Results in an output file `Cell_type_centrality_summary.txt` that documents calculated centralities and properties of the intercellular disease model.}
\usage{
NicheNet_cell_type_centrality(all_ligand_activity, out_dir = "")
}
\arguments{
\item{all_ligand_activity}{Matrix. Outcome of `NicheNet_ligand_activity_analysis()`.}

\item{out_dir}{String. Indicates the full path to the folder in which results should be stored.}
}
\value{
Returns a matrix that summarizes cell type centralities in the multicellular disease module (MCDM)
}
\description{
NicheNet ligand activity analysis for all cell types
Will test cell type centrality based on NicheNet ligand activities. Results in an output file `Cell_type_centrality_summary.txt` that documents calculated centralities and properties of the intercellular disease model.
}
