% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheNet_ligand_activity_analysis.R
\name{NicheNet_ligand_activity_analysis}
\alias{NicheNet_ligand_activity_analysis}
\title{NicheNet ligand activity analysis for all cell types
Will test NicheNet ligand activity for all possible cell type interactions (inclduign self-interactions). Genes are formatted as Entrez Symbols in order to be compatible with the standard ligand target interactions and ligand receptor network.}
\usage{
NicheNet_ligand_activity_analysis(
  degs,
  background_genes,
  out_dir = "",
  ligand_target = NULL,
  lr_network = NULL,
  cores = 1
)
}
\arguments{
\item{degs}{Matrix. Every column represents one cell types DEGs. Column names indicate the cell type.}

\item{background_genes}{Matrix. Every column represents one cell types background genes. Same column names as in `degs`. Background genes for cell types can be derived from `background_genes_NicheNet()`}

\item{out_dir}{String. Indicates the full path to the folder in which NicheNet ligand activity meta-data and results should be stored.}

\item{ligand_target}{Matrix. Targets in rows, ligands in columns. If NULL, the standard information from NicheNet will be downloaded: https://zenodo.org/record/3260758/files/ligand_target_matrix.rds}

\item{lr_network}{Matrix. If NULL, the standard information from NicheNet will be downloaded: https://zenodo.org/record/3260758/files/lr_network.rds}

\item{cores}{Integer. Number of cores to be used in parallel computing. Defaults to number of cores available - 1.}
}
\value{
Returns a matrix including all cell type's ligand activities.
}
\description{
NicheNet ligand activity analysis for all cell types
Will test NicheNet ligand activity for all possible cell type interactions (inclduign self-interactions). Genes are formatted as Entrez Symbols in order to be compatible with the standard ligand target interactions and ligand receptor network.
}
