% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_effect_on_fold_change.R
\name{pharma_effect_on_fold_change}
\alias{pharma_effect_on_fold_change}
\title{Pharmacological action on targeted DEGs}
\usage{
pharma_effect_on_fold_change(
  drug_dist_files,
  deg_files,
  pharma_effect,
  save_names,
  out_dir = "",
  cores = 1
)
}
\arguments{
\item{drug_dist_files}{Vector with complete filepaths to the average closest network distance files (generated by `average_closest_distance_network_drug_screening()`) for which the pharmacological action on targeted DEGs shall be summarized}

\item{deg_files}{Vector with the same length drug_dist_files with complete filepaths to the DEG files (generated by `calculate_DEGs()` in this example). Every element in `deg_files` is matched with the corresponding element in `drug_dist_files`.}

\item{pharma_effect}{Matrix with three columns, in which the first column specifies the drug identifier matchable with the first column in the average closest netwrok distance files. The second column specifies the drug target and column three specifies the pharmacological action on the drug target.}

\item{save_names}{Vector specifying the names for output. Each file corresponds to the processed version of a `drug_dist_files`.}

\item{out_dir}{Character string, indicating the directory in which output shall be saved.}

\item{cores}{Integer specifying the number of cores to be used in computation.}
}
\description{
Used for manual evaluation of pharmacological counteraction of DEG fold change
}
