% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppin_formatting.R
\name{ppin_formatting}
\alias{ppin_formatting}
\title{Protein-protein interaction network (PPIN) formatting}
\usage{
ppin_formatting(ppin)
}
\arguments{
\item{ppin}{matrix with 2 columns specifying all protein-protein interactions in the PPIN}
}
\value{
A 2 column matrix specifying the protein-protein interactions included in the PPINs largest connected component
}
\description{
Protein-protein interaction network (PPIN) formatting
}
\examples{
load_example_data()
ppin_formatting(test_ppi)

}
