% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_drug_target_matrix_for_network_distance_calculation.R
\name{prepare_drug_target_matrix_for_network_distance_calculation}
\alias{prepare_drug_target_matrix_for_network_distance_calculation}
\title{Preparation of drug target matrix for network distance analysis
Removes all drug targets not found in PPIN and thereafter identifies unique drug target combinations.
Unique drug target combinations are then used for calculation of network distances to save computational time and
can thereafter be translated back to individual drugs.}
\usage{
prepare_drug_target_matrix_for_network_distance_calculation(
  ppin,
  drug_target_matrix,
  file_name,
  out_dir
)
}
\arguments{
\item{ppin}{Matrix with 2 columns, every row represents one protein-protein interaction in the protein-protein interaction network (PPIN) largest connected component (LCC). Output from ppin_formatting()}

\item{drug_target_matrix}{Matrix listing each drug's targets in a column.}

\item{file_name}{Character string specifying filename for the output.}

\item{out_dir}{Character string specifying path to out directory where data will be saved.}
}
\value{
Returns drug_taregt_matrix including only unique drug target combinations. The matching between unique drug target combinations and
 their respective drugs are documented in paste(out_dir,"/SAME_DRUG_TARGETS_",file_name,".txt",sep="")
}
\description{
Preparation of drug target matrix for network distance analysis
Removes all drug targets not found in PPIN and thereafter identifies unique drug target combinations.
Unique drug target combinations are then used for calculation of network distances to save computational time and
can thereafter be translated back to individual drugs.
}
