% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_setup_for_v_3_6_3.R
\name{R_setup_for_v_3_6_3}
\alias{R_setup_for_v_3_6_3}
\title{Set up R environment for R 3.6.3}
\usage{
R_setup_for_v_3_6_3()
}
\value{
Loads all required packages for R v3.6.3
}
\description{
Set up R environment for R 3.6.3
}
\examples{
R_setup_for_v_3.6.3()
}
