% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA_drugs.R
\docType{data}
\name{RA_drugs}
\alias{RA_drugs}
\title{List of approved RA drugs, according to DrugBank}
\format{
A vector with DrugBank ID as values and drug name as names.
}
\usage{
RA_drugs
}
\description{
To identify drugs that are approved for the treatment of a disease, 
DrugBank was screened according to the "Indication" category. 
More details can be found in the supplement of the publication.
}
\keyword{datasets}
