% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_drug_target_bin_adjusted_distances.R
\name{random_drug_target_bin_adjusted_distances}
\alias{random_drug_target_bin_adjusted_distances}
\title{Calculation of closest distances between node degree adjusted, randomized drug target combinations and node degree adjusted, randomized disease genes for creation of a reference distribution}
\usage{
random_drug_target_bin_adjusted_distances(
  bins,
  drug_target_matrix,
  disease_genes,
  ppin_dist,
  seed
)
}
\arguments{
\item{bins}{bins used for network degree adjustment as previously obtained using bin_creation_by_min_bin_size()}

\item{drug_target_matrix}{Matrix listing each unique drug target combination in a column.}

\item{disease_genes}{Vector specifying disease associated genes in the PPIN.}

\item{ppin_dist}{Distances between all proteins included in the LCC of the PPIN.}

\item{seed}{Seed used for randomization of drug targets and disease genes. Needs to be altered for every iteration.}
}
\value{
Returns a vector containing average closest distances between every set of random (bin-adjusted) disease genes and random (bin-adjusted) drug targets.
}
\description{
Calculation of closest distances between node degree adjusted, randomized drug target combinations and node degree adjusted, randomized disease genes for creation of a reference distribution
}
\keyword{internal}
