% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/seperate_unique_drug_target_combinations_into_individual_drugs.R
\name{seperate_unique_drug_target_combinations_into_individual_drugs}
\alias{seperate_unique_drug_target_combinations_into_individual_drugs}
\title{Seperation of unique drug targets to individual drugs.}
\usage{
seperate_unique_drug_target_combinations_into_individual_drugs(
  files,
  same_drugs,
  disease_specific_drugs,
  in_dir = getwd(),
  out_dir = getwd(),
  output_file_name_add_on = "INDIVIDUAL_DRUGS_",
 
    remove_part_of_file_name_for_output = "drug-disease_closest_distances_vs_random_bin_adjusted\\\\__",
  cores = 1
)
}
\arguments{
\item{files}{character vector; specifying file names for ```average_closest_distance_network_screening()``` output files that this function shall be applied to.}

\item{same_drugs}{matrix with 2 columns; specifying mapping of individual drugs to unique drug target combinations created by ```prepare_drug_target_matrix_for_network_distance_calculation()```. First column indicating the DrugBankID used for the unique drug target combination in average_closest_distance_network_screening() and second column indicating the mapping to individual drugs.}

\item{disease_specific_drugs}{character vector; specifying DrugBankID of disease specific drugs (drug names for known drug-disease pairs).}

\item{in_dir}{character string; specifying path to directory of 'files'. Too long directory paths can cause problems for ```read.table()``` in this function (at least on Windows).}

\item{out_dir}{character string; specifying path to directory for the output files}

\item{output_file_name_add_on}{character string; file name add on for output files. Set to "" if you do not wish to change filenames. If out_dir = in_dir and output_file_name_add_on = "" this function will overwrite 'files'.}

\item{remove_part_of_file_name_for_output}{character string; part of file name in 'files' that is to be removed from file name of output file.}

\item{cores}{integer; number of cores used for parallel computing.}
}
\value{
Creates a number of files equal to length(files) in out_dir that includes network distances for individual drugs instead of unique drug target combinations.
}
\description{
Seperation of unique drug targets to individual drugs.
}
\examples{
same_drugs = as.matrix(read.table(file = "data/SAME_DRUG_TARGETS_in_lit_ppin.txt", sep="\t", header = T, stringsAsFactors = F))
disease_specific_drugs = as.matrix(read.table(file = "data-raw/RA_drugs_from_DrugBank_200214.txt",sep="\t", header = T, stringsAsFactors = F))
seperate_unique_drug_target_combinations_into_individual_drugs(files = "data/test.txt", same_drugs = same_drugs, disease_specific_drugs = disease_specific_drugs)

}
