% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_network_proximity_calculation.R
\name{summary_network_proximity_calculation}
\alias{summary_network_proximity_calculation}
\title{Evaluation of network screening}
\usage{
summary_network_proximity_calculation(
  files,
  deg_files,
  ppin,
  top_degs = "all",
  drug_effect_data,
  file_names,
  transl = NULL,
  out_dir = getwd(),
  cores = 1
)
}
\arguments{
\item{files}{character vector; full path and file names for files from ```seperate_unique_drug_target_combinations_into_individual_drugs()```.}

\item{deg_files}{character vector same length as `files`; specifying full path and file name for the DEG file that should be used for a given element in `files`.}

\item{ppin}{matrix with 2 columns representing all interactions in the protein-protein interaction network (PPIN) used for network calculations.}

\item{top_degs}{either integer vector of same length as 'files' specifying the number of top DEGs used for calculation of network distances or a character string  specifying "all", in case all DEGs were used. If 'top_DEGs' is supplied as a numeric vector, a value in top_DEGs that is higher than the available number of genes in 'deg_files' will result in the use of all DEGs.}

\item{drug_effect_data}{matrix, first column = DrugBank ID, second = drug name, third = drug target, fourth = pharmacological effect on target.}

\item{file_names}{vector same length as `files`; specifying output file names without file path. If `file_names` is equal to 'files' and `out_dir` is getwd() this function will override network distance calculation files.}

\item{transl}{matrix; 3 columns. Column 1 matches gene annotation in `ppin`, column 2 matches gene annotation in `deg_files`, column 3 matches gene annotation in `DrugBank_info`. Not required if genes are annotated similarly.}

\item{out_dir}{character string; specifying output directory.}

\item{cores}{number of cores used for computing.}
}
\description{
Creation of summary files used for evaluation of drug effect on the targeted genes.
Counteraction of differential expressed genes by drug targets is later used for drug candidate selection.
}
